requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-http==4.5.0
pandas==2.1.0
requests==2.31.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
